Refactored project code
